Refactor namespace for generic compatibility
