+++ examples/baseball.py
@@ -12,7 +12,7 @@ import numpyro
 import numpyro.distributions as dist
 from numpyro.examples.datasets import BASEBALL, load_dataset
 from numpyro.infer_util import log_likelihood, predictive
-from numpyro.mcmc import MCMC, NUTS
+from numpyro.infer import MCMC, NUTS
 
 
 """
