+++ examples/neutra.py
@@ -20,8 +20,8 @@ import numpyro.distributions as dist
 from numpyro.distributions import constraints
 from numpyro.hmc_util import initialize_model
 from numpyro.infer_util import transformed_potential_energy
-from numpyro.mcmc import MCMC, NUTS
-from numpyro.svi import SVI, elbo
+from numpyro.infer import MCMC, NUTS
+from numpyro.infer import SVI, elbo
 
 # TODO: remove when the issue https://github.com/google/jax/issues/939 is fixed upstream
 # The behaviour when training guide under fast math mode is unstable.
