+++ numpyro/compat/distributions.py
@@ -1,2 +1,4 @@
 from numpyro.distributions import *  # noqa: F401, F403
-from numpyro.distributions.constraints import *  # noqa: F401, F403
+import numpyro.distributions.constraints as constraints  # noqa: F401, F403
+
+transforms = constraints
