+++ numpyro/compat/ops.py
@@ -1,4 +1,3 @@
 from jax.numpy import *  # noqa: F401, F403
 
-
 tensor = array  # noqa: F405
