+++ numpyro/handlers.py
@@ -81,7 +81,8 @@ from collections import OrderedDict
 from jax import random
 import jax.numpy as np
 
-from numpyro.distributions.constraints import ComposeTransform, biject_to, real
+from numpyro.distributions.constraints import real
+from numpyro.distributions.transforms import ComposeTransform, biject_to
 from numpyro.primitives import Messenger
 
 __all__ = [
