+++ numpyro/infer_util.py
@@ -8,7 +8,8 @@ from jax.tree_util import tree_flatten
 
 import numpyro
 import numpyro.distributions as dist
-from numpyro.distributions.constraints import ComposeTransform, biject_to, real
+from numpyro.distributions.constraints import real
+from numpyro.distributions.transforms import ComposeTransform, biject_to
 from numpyro.handlers import block, condition, seed, substitute, trace
 from numpyro.util import while_loop
